finally Backend Completed

+++ src/main/java/com/SmartTaskAndProjectManagementSystem/Entity/TeamMember.java
@@ -18,8 +18,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
-
 @Entity 
 @Data
 @Builder
@@ -32,7 +30,7 @@ public class TeamMember extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 
 	@Column(nullable = false)
 	private String name;
